requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.0
drf-spectacular>=0.26.2
psycopg[binary]>=3.1
celery>=5.3
redis>=4.5
